# frontend/person-blog-app/packages/blog/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Gateway.API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/video"
    }
  ],
  "paths": {
...
    "/api/Blog/blog/{blogId}": {
      "get": {
        "tags": [
          "Blog"
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlogModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Channel/{channelId}": {
      "get": {
        "tags": [
          "Channel"
        ],
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Channel/posts/{channelId}": {
      "get": {
        "tags": [
          "Channel"
        ],
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostCommonModelV2PagedListViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCommonModelV2PagedListViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCommonModelV2PagedListViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Channel/playLists/{channelId}": {
      "get": {
        "tags": [
          "Channel"
        ],
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Subscriber/subscribe/{blogId}": {
      "post": {
        "tags": [
          "Subscriber"
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Subscriber/unsubscribe/{blogId}": {
      "post": {
        "tags": [
          "Subscriber"
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Subscriber/subscriptions": {
      "get": {
        "tags": [
          "Subscriber"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/VideoUpload/initiate": {
      "post": {
        "tags": [
          "VideoUpload"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/VideoUpload/generate-url": {
      "post": {
        "tags": [
          "VideoUpload"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateUrlRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateUrlRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateUrlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PreSignedUrl"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreSignedUrl"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreSignedUrl"
                }
              }
            }
          }
        }
      }
    },
    "/api/VideoUpload/complete": {
      "post": {
        "tags": [
          "VideoUpload"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/VideoUpload/abort": {
      "post": {
        "tags": [
          "VideoUpload"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortUploadRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortUploadRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AbortUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/VideoUpload/session/{uploadId}": {
      "get": {
        "tags": [
          "VideoUpload"
        ],
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/VideoUpload/parts/{uploadId}": {
      "get": {
        "tags": [
          "VideoUpload"
        ],
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MultipartUploadPart"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MultipartUploadPart"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MultipartUploadPart"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbortUploadRequest": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BlogModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "subscribersCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CompleteUploadRequest": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "nullable": true
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MultipartUploadPart"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GenerateUrlRequest": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "nullable": true
          },
          "partNumber": {
            "type": "integer",
            "format": "int32"
          },
          "expiryMinutes": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "InitiateUploadRequest": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "string",
            "format": "uuid"
          },
          "objectName": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "fileExtension": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "MultipartUploadPart": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "MultipartUploadSession": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "nullable": true
          },
          "bucketId": {
            "type": "string",
            "nullable": true
          },
          "objectName": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/UploadStatus"
          },
          "totalSize": {
            "type": "integer",
            "format": "int64"
          },
          "totalParts": {
            "type": "integer",
            "format": "int32"
          },
          "completedETag": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PostCommonModelV2": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "blogId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "postType": {
            "$ref": "#/components/schemas/PostTypeModel"
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PostCommonModelV2PagedListViewModel": {
        "required": [
          "items",
          "pageSize",
          "totalPageCount",
          "totalPostsCount"
        ],
        "type": "object",
        "properties": {
          "totalPageCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPostsCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostCommonModelV2"
            }
          }
        },
        "additionalProperties": false
      },
...
      "PostTypeModel": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "PreSignedUrl": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "nullable": true
          },
          "uploadId": {
            "type": "string",
            "nullable": true
          },
          "partNumber": {
            "type": "integer",
            "format": "int32"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "httpMethod": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UploadStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  },
...
}